+++ SWdemo/src/Iservice.java
@@ -1,9 +1,9 @@
 public interface Iservice {
     String name=null;
-    int discount_amoutn=0;
+    double discount_amoutn=0;
     String getName();
-    void setDiscount_amoutn(int discount_amoutn);
-    int GetDiscount();
+    void setDiscount_amoutn(double discount_amoutn);
+    double GetDiscount();
 
 
 }
