+++ SWdemo/src/LandLine.java
@@ -1,6 +1,6 @@
 public class LandLine implements Iservice{
     String name="Landline";
-    int discount_amoutn=0;
+    double discount_amoutn=0;
 
 
     @Override
@@ -8,12 +8,12 @@ public class LandLine implements Iservice{
         return name;
     }
     @Override
-    public void setDiscount_amoutn(int discount_amoutn) {
+    public void setDiscount_amoutn(double discount_amoutn) {
         this.discount_amoutn=discount_amoutn;
     }
 
     @Override
-    public int GetDiscount() {
+    public double GetDiscount() {
         return this.discount_amoutn;
     }
